kleine aanpassing bij AanpassenMenu().

+++ RestaurantAppB/Classes/Gerecht.cs
@@ -9,7 +9,7 @@ namespace RestaurantApp.Classes
         public string gang { get; set; }
         public string Naam { get; set; }
         public int GerechtNummer { get; set; }
-        public int Prijs { get; set; }
+        public double Prijs { get; set; }
         public string Ingredienten { get; set; }
     }
 }
